+++ packages/elasticsearch-store/src/index.ts
@@ -1,11 +1,12 @@
 import Cluster from './cluster';
 import IndexManager from './index-manager';
 import IndexStore from './index-store';
+import { timeseriesIndex } from './utils';
 
 export * from './interfaces';
-export * from './utils';
 export {
     Cluster,
     IndexManager,
     IndexStore,
+    timeseriesIndex,
 };
